requirements.txt
langchain
openai
python-dotenv
slack-sdk 
slack-bolt 
flask
gunicorn
urllib3<2